+++ CHANGELOG.md
@@ -15,10 +15,12 @@ v0.6.0
  - `Exponential` now uses the Ziggurat methd to generate random samples.
  - `Binomial` now implements `Univariate<u64, f64>` rather than `Univariate<i64, f64>`, meaning `Binomial::min` and `Binomial::max` now return `u64`
  - `Bernoulli` now implements `Univariate<u64, f64>` rather than `Univariate<i64, f64>`, meaning `Bernoulli::min` and `Bernoulli::min` now return `u64`
+ - `Geometric` now implements `Univariate<u64, f64>` rather than `Univariate<i64, f64>`, meaning `Geometric::min` and `Geometric::min` now return `u64`
  - `Binomial` now implements `Mode<u64>` instead of `Mode<i64>`
  - `Bernoulli` now implements `Mode<u64>` instead of `Mode<i64`
  - `Binomial` now implements `Discrete<u64, f64>` rather than `Discrete<i64, f64>`
  - `Bernoulli` now implements `Discrete<u64, f64>` rather than `Discrete<i64, f64>`
+ - `Geometric` now implements `Discrete<u64, f64>` rather than `Discrete<i64, f64>`
 
 v0.5.1
  - Fixed critical bug in `normal::sample_unchecked` where it was returning `NaN`